chore: 안쓰는 코드/파일 정리 및 폴더 구조 정비
