--- Kartrider/Feature/Common/ContentView.swift
@@ -1,22 +1,0 @@
-//
-//  ContentView.swift
-//  Kartrider
-//
-//  Created by jiwon on 5/27/25.
-//
-
-import SwiftData
-import SwiftUI
-
-// TODO: 안쓰는 파일 및 코드 삭제
-struct ContentView: View {
-
-    var body: some View {
-        Text("네비게이션 스택 옮기기 완료")
-    }
-}
-
-#Preview {
-    ContentView()
-    //        .modelContainer(for: Item.self, inMemory: true)
-}
