+++ Kartrider/Feature/Outro/OutroViewModel.swift
