--- Kartrider/Feature/Storage/ViewModel/StorageViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  StorageViewModel.swift
-//  Kartrider
-//
-//  Created by J on 5/31/25.
-//
-
-import Foundation
