+++ Kartrider/Repository/DTO/StoryJSON.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 
-// MARK: - 파일 이름 바꾸세요~
 struct StoryJSON: Decodable {
     let meta: MetaData
     let story: StoryData
