--- Kartrider/Utility/Font.swift
@@ -1,9 +1,0 @@
-//
-//  Font.swift
-//  Kartrider
-//
-//  Created by jiwon on 5/27/25.
-//
-
-// TODO: ?
-import Foundation
