--- Pickacha Watch App/Stage.swift
@@ -1,15 +1,0 @@
-////
-////  Stage.swift
-////  Pickacha Watch App
-////
-////  Created by jiwon on 6/4/25.
-////
-//
-//import Foundation
-//
-//enum Stage: String {
-//    case idle
-//    case exposition
-//    case decision
-//    case ending
-//}
