--- Pickacha Watch App/WatchFeature/Common/ContentView.swift
@@ -1,22 +1,0 @@
-//
-//  ContentView.swift
-//  Pickacha Watch App
-//
-//  Created by jiwon on 5/28/25.
-//
-
-// TODO: 삭제!!!!!!!
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Text("네비게이션 스택 옮기기 완료")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
